backend/requirements.txt
flask>=3.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
flask-cors>=4.0.0
requests-toolbelt>=1.0.0
